+++ templates/single-event.php
@@ -3,8 +3,8 @@
  * The Template for displaying the single event posts.
  *
  * @package WordPress
- * @subpackage %NAME%
- * @since %NAME% 1.0.0
+ * @subpackage Wolf Events
+ * @since Wolf Events 1.0.0
  */
 get_header( 'events' );
 ?>