+++ CHANGELOG.md
@@ -1,3 +1,43 @@
+#### 15th June 2010 - Version 1.2.1
+
+* Tweak: Added custom date class filter
+
+#### 31th October 2019 - Version 1.2.0
+
+* Tweak: Added Bandsintown button option
+
+#### 6th August 2019 - Version 1.1.9
+
+* Fix: Updater date issue
+
+#### 8th July 2019 - Version 1.1.8
+
+* Tweak: Updater for Wolf Tour Dates now import post slug as well (for Decibel v3)
+
+#### 6th June 2019 - Version 1.1.7
+
+* Tweak: Added updater for Wolf Tour Dates (for Decibel v3)
+
+#### 2nd March 2019 - Version 1.1.6
+
+* Tweak: Added currency metabox for microdata
+
+#### 13th January 2019 - Version 1.1.5
+
+* Fix: Changed raw date format to ISO
+
+#### 2nd December 2018 - Version 1.1.4
+
+* Fix: Taxonomy template loader
+
+#### 20th August 2018 - Version 1.1.3
+
+* Tweak: Added we_ticket_link_target filter
+
+#### 13th July 2018 - Version 1.1.2
+
+* Fix: Date translation issue
+
 #### 7th May 2018 - Version 1.1.1
 
 * Tweak: Add post state in admin